update scripts from skeleton project, fix style check warnings, remove langauge example code

+++ clear-cache.sh
@@ -1,18 +1,15 @@
 #!/bin/sh -e
 
-SCRIPT_DIR=$(cd $(dirname ${0}); pwd)
-
-echo "Deleting python cache files and directories and the .egg-info directory."
-
-FILES="build .pyvenv .coverage .sonar"
+echo "Deleting cached and generated files."
+DIR=$(dirname "${0}")
+SCRIPT_DIR=$(cd "${DIR}"; pwd)
+FILES="build .pyvenv .coverage"
 
 for FILE in ${FILES}; do
     if [ -e "${FILE}" ]; then
         echo "rm -rf ${FILE}"
-        rm -rf "${SCRIPT_DIR}/${FILE}"
+        rm -rf "${SCRIPT_DIR:?}/${FILE}"
     fi
 done
 
-find "${SCRIPT_DIR}" -name "__pycache__" | xargs -I {} rm -rfv "{}"
-find "${SCRIPT_DIR}" -name "*.egg-info" | xargs -I {} rm -rfv "{}"
-find "${SCRIPT_DIR}" -name "*.pyc" | xargs -I {} rm -fv "{}"
+find "${SCRIPT_DIR}" \( -name '__pycache__' -o -name '*.egg-info' -o -name '*.pyc' \) -exec rm -rf {} +
